itens e diagramas

+++ src/player/IPlayer.java
@@ -2,7 +2,7 @@ package lccr.player.ifaces;
 
 public interface IPlayer extends IFighter extends IHandler extends IWanderer
 {
-	public double getCR();
-	public double getKnowledge();
-	public double getMigue();
+	public float getCR();
+	public float getKnowledge();
+	public float getMigue();
 }
